// frontend/src/app/api/model/planCreate.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanItemRef } from './planItemRef';


export interface PlanCreate { 
    title: string;
    /**
     * Existing, distinct tasks/habits/learning cards (BR-10)
     */
    items: Array<PlanItemRef>;
    estimatedDurationMinutes: number;
    startDateTime: string;
    /**
     * Must be after startDateTime (BR-11)
     */
    endDateTime: string;
    /**
     * 1 is the highest
     */
    priorityOrder: number;
}