// frontend/src/app/api/model/planItemRef.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanSourceType } from './planSourceType';


export interface PlanItemRef { 
    sourceType: PlanSourceType;
    sourceId: number;
}
export namespace PlanItemRef {
}